Se agrega tabla measures y se modifican vistas para tener todo de medidas y clientes y proveedores se vuelven nullables los campos rfc,direccion, teléfono, correo. Se actualizan los detalles venta y compra para que envíen información adicional, así como los métodos get_product_by_id para que se envíe la relación con measures. Doy acceso para que puedan dar de alta usuarios sin que aparezca el rol de superAdmin.

+++ database/migrations/2021_04_01_190201_create_purchase_details_table.php
@@ -15,7 +15,7 @@ class CreatePurchaseDetailsTable extends Migration
     {
         Schema::create('purchase_details', function (Blueprint $table) {
             $table->id();
-            $table->integer('quantity');
+            $table->decimal('quantity', 12, 3);
             $table->decimal('price');
             $table->unsignedBigInteger('purchase_id');
             $table->foreign('purchase_id')->references('id')->on('purchases');
